Example

Example Julia package repo.

Metatheory.BackoffScheduler — Type

A Rewrite Scheduler that implements exponential rule backoff. For each rewrite, there exists a configurable initial match limit. If a rewrite search yield more than this limit, then we ban this rule for number of iterations, double its limit, and double the time it will be banned next time.

This seems effective at preventing explosive rules like associativity from taking an unfair amount of resources.

Metatheory.RhsFunCache — Type

Type representing a cache of RuntimeGeneratedFunctions corresponding to right hand sides of dynamic rules

Metatheory.amp — Method

Add a & expression

Metatheory.block — Method

Make a block expression from an array of exprs

Metatheory.cleanast — Method

Binarize n-ary operators (+ and *) and call rmlines

Metatheory.dollar — Method

Add a dollar expression

Metatheory.eval_types_in_assertions — Method

When creating a theory, type assertions in the left hand contain symbols. We want to replace the type symbols with the real type values, to fully support the subtyping mechanism during pattern matching.

Metatheory.genrhsfun — Method

Generates a tuple containing the list of formal parameters (Symbols) and the RuntimeGeneratedFunction corresponding to the right hand side of a :dynamic Rule.

Metatheory.rmlines — Method

Remove LineNumberNode from quoted blocks of code