Metatheory

Metatheory.RhsFunCache — Type

Type representing a cache of RuntimeGeneratedFunctions corresponding to right hand sides of dynamic rules

Metatheory.eval_types_in_assertions — Method

When creating a theory, type assertions in the left hand contain symbols. We want to replace the type symbols with the real type values, to fully support the subtyping mechanism during pattern matching.

Metatheory.genrhsfun — Method

Generates a tuple containing the list of formal parameters (Symbols) and the RuntimeGeneratedFunction corresponding to the right hand side of a :dynamic Rule.