Metatheory.Util

Metatheory.Util — Module

Definitions of various utility functions for metaprogramming

Metatheory.Util.amp — Method

Add a & expression

Metatheory.Util.bf_walk! — Method

Breadth First Walk (Tree Prewalk) on expressions mutates expression in-place.

Metatheory.Util.bf_walk — Method

Breadth First Walk (Tree Prewalk) on expressions. Does not mutate expressions.

Metatheory.Util.binarize! — Method

HARD FIX of n-arity of operators in Expr trees

Metatheory.Util.block — Method

Make a block expression from an array of exprs

Metatheory.Util.cleanast — Method

Binarize n-ary operators (+ and *) and call rmlines

Metatheory.Util.df_walk! — Method

Depth First Walk (Tree Postwalk) on expressions, mutates expression in-place.

Metatheory.Util.df_walk — Method

Depth First Walk (Tree Postwalk) on expressions. Does not mutate expressions.

Metatheory.Util.dollar — Method

Add a dollar expression

Metatheory.Util.normalize — Method

Iterates a function f on datum until a fixed point is reached where f(x) == x

Metatheory.Util.normalize_nocycle — Method

Like normalize but keeps a vector of hashes to detect cycles, returns the current datum when a cycle is detected

Metatheory.Util.rmlines — Method

Remove LineNumberNode from quoted blocks of code