Scheduling for E-Graph rewriting

Metatheory.EGraphs.Schedulers.AbstractScheduler — Type

abstract type AbstractScheduler

Represents a rule scheduler for the equality saturation process

Fields

Metatheory.EGraphs.Schedulers.BackoffScheduler — Type

struct BackoffScheduler <: Metatheory.EGraphs.Schedulers.AbstractScheduler

A Rewrite Scheduler that implements exponential rule backoff. For each rewrite, there exists a configurable initial match limit. If a rewrite search yield more than this limit, then we ban this rule for number of iterations, double its limit, and double the time it will be banned next time.

This seems effective at preventing explosive rules like associativity from taking an unfair amount of resources.

Fields

data::Dict{Rule,Metatheory.EGraphs.Schedulers.BackoffSchedulerEntry}

G::EGraph

theory::Array{var"#s66",1} where var"#s66"<:Rule

Metatheory.EGraphs.Schedulers.ScoredScheduler — Type

struct ScoredScheduler <: Metatheory.EGraphs.Schedulers.AbstractScheduler

A Rewrite Scheduler that implements exponential rule backoff. For each rewrite, there exists a configurable initial match limit. If a rewrite search yield more than this limit, then we ban this rule for number of iterations, double its limit, and double the time it will be banned next time.

This seems effective at preventing explosive rules like associativity from taking an unfair amount of resources.

Fields

data::Dict{Rule,Metatheory.EGraphs.Schedulers.ScoredSchedulerEntry}

G::EGraph

theory::Array{var"#s66",1} where var"#s66"<:Rule

Metatheory.EGraphs.Schedulers.SimpleScheduler — Type

struct SimpleScheduler <: Metatheory.EGraphs.Schedulers.AbstractScheduler

A simple Rewrite Scheduler that applies every rule every time

Fields

Metatheory.EGraphs.Schedulers.cansaturate — Method

Should return true if the e-graph can be said to be saturated

Signatures

cansaturate(s::Metatheory.EGraphs.Schedulers.AbstractScheduler) -> Bool

Methods

cansaturate(s)

defined at /home/runner/work/Metatheory.jl/Metatheory.jl/src/EGraphs/Schedulers/abstractscheduler.jl:12.

cansaturate(s)

defined at /home/runner/work/Metatheory.jl/Metatheory.jl/src/EGraphs/Schedulers/backoffscheduler.jl:48.

cansaturate(s)

defined at /home/runner/work/Metatheory.jl/Metatheory.jl/src/EGraphs/Schedulers/scoredscheduler.jl:78.

cansaturate(s)

defined at /home/runner/work/Metatheory.jl/Metatheory.jl/src/EGraphs/Schedulers/simplescheduler.jl:6.

Metatheory.EGraphs.Schedulers.readstep! — Method

This function is called before pattern matching on the e-graph

Signatures

readstep!(s::Metatheory.EGraphs.Schedulers.AbstractScheduler)

Methods

readstep!(s)

defined at /home/runner/work/Metatheory.jl/Metatheory.jl/src/EGraphs/Schedulers/abstractscheduler.jl:22.

readstep!(s)

defined at /home/runner/work/Metatheory.jl/Metatheory.jl/src/EGraphs/Schedulers/backoffscheduler.jl:50.

readstep!(s)

defined at /home/runner/work/Metatheory.jl/Metatheory.jl/src/EGraphs/Schedulers/scoredscheduler.jl:80.

readstep!(s)

defined at /home/runner/work/Metatheory.jl/Metatheory.jl/src/EGraphs/Schedulers/simplescheduler.jl:13.

Metatheory.EGraphs.Schedulers.shouldskip — Method

Should return true if the rule r should be skipped

Signatures

shouldskip(s::Metatheory.EGraphs.Schedulers.AbstractScheduler, r::Rule) -> Bool

Methods

shouldskip(s, r)

defined at /home/runner/work/Metatheory.jl/Metatheory.jl/src/EGraphs/Schedulers/abstractscheduler.jl:17.

shouldskip(s, r)

defined at /home/runner/work/Metatheory.jl/Metatheory.jl/src/EGraphs/Schedulers/backoffscheduler.jl:27.

shouldskip(s, r)

defined at /home/runner/work/Metatheory.jl/Metatheory.jl/src/EGraphs/Schedulers/scoredscheduler.jl:28.

shouldskip(s, r)

defined at /home/runner/work/Metatheory.jl/Metatheory.jl/src/EGraphs/Schedulers/simplescheduler.jl:7.

Metatheory.EGraphs.Schedulers.writestep! — Method

This function is called after pattern matching on the e-graph

Signatures

writestep!(s::Metatheory.EGraphs.Schedulers.AbstractScheduler, r::Rule) -> Union{Nothing, Int64}

Methods

writestep!(s, r)

defined at /home/runner/work/Metatheory.jl/Metatheory.jl/src/EGraphs/Schedulers/abstractscheduler.jl:27.

writestep!(s, rule)

defined at /home/runner/work/Metatheory.jl/Metatheory.jl/src/EGraphs/Schedulers/backoffscheduler.jl:66.

writestep!(s, rule)

defined at /home/runner/work/Metatheory.jl/Metatheory.jl/src/EGraphs/Schedulers/scoredscheduler.jl:96.

writestep!(s, r)

defined at /home/runner/work/Metatheory.jl/Metatheory.jl/src/EGraphs/Schedulers/simplescheduler.jl:14.