API Documentation

Metatheory

Main.Metatheory.Rule — Type

mutable struct Rule

Structure representing rewrite rules.

Fields

left::Any

right::Any

expr::Expr

mode::Symbol

Main.Metatheory.compile_theory — Method

Compile a theory to a closure that does the pattern matching job Returns a RuntimeGeneratedFunction, which does not use eval and is as fast as a regular Julia anonymous function 🔥

Signatures

compile_theory(theory::Array{Main.Metatheory.Rule,1}, mod::Module; __source__) -> RuntimeGeneratedFunctions.RuntimeGeneratedFunction{_A,_B,_C,_D} where _D where _C where _B where _A

Methods

compile_theory(theory, mod; __source__)

defined at /home/runner/work/Metatheory.jl/Metatheory.jl/src/matchcore_compiler.jl:61.

Main.Metatheory.@compile_theory — Macro

Compile a theory at runtime to a closure that does the pattern matching job

Signatures

Methods

Metatheory.EGraphs

Main.Metatheory.EGraphs.AbstractAnalysis — Type

abstract type AbstractAnalysis

Fields

Main.Metatheory.EGraphs.EGraph — Type

mutable struct EGraph

Fields

U::DataStructures.IntDisjointSets

equality relation over e-class ids

M::Dict{Int64,Array{Any,1}}

H::Dict{Any,Int64}

parents::Dict{Int64,Array{Tuple{Any,Int64},1}}

dirty::Array{Int64,1}

root::Int64

analyses::Array{Main.Metatheory.EGraphs.AbstractAnalysis,1}

Main.Metatheory.EGraphs.ExtractionAnalysis — Type

struct ExtractionAnalysis <: Main.Metatheory.EGraphs.AbstractAnalysis

An AbstractAnalysis that computes the cost of expression nodes and chooses the node with the smallest cost for each E-Class.

Fields

egraph::Main.Metatheory.EGraphs.EGraph

costfun::Function

data::Dict{Int64,Tuple{Any,Number}}

Main.Metatheory.EGraphs.RhsFunCache — Type

Type representing a cache of RuntimeGeneratedFunctions corresponding to right hand sides of dynamic rules

Main.Metatheory.EGraphs.astsize — Method

A basic cost function, where the computed cost is the size (number of children) of the current expression.

Signatures

astsize(n::Any) -> Int64

Methods

astsize(n)

defined at /home/runner/work/Metatheory.jl/Metatheory.jl/src/EGraphs/extraction.jl:5.

astsize(n)

defined at /home/runner/work/Metatheory.jl/Metatheory.jl/src/EGraphs/extraction.jl:6.

Main.Metatheory.EGraphs.astsize_inv — Method

A basic cost function, where the computed cost is the size (number of children) of the current expression, times -1. Strives to get the largest expression

Signatures

astsize_inv(n::Any) -> Int64

Methods

astsize_inv(n)

defined at /home/runner/work/Metatheory.jl/Metatheory.jl/src/EGraphs/extraction.jl:13.

astsize_inv(n)

defined at /home/runner/work/Metatheory.jl/Metatheory.jl/src/EGraphs/extraction.jl:14.

Main.Metatheory.EGraphs.ematchlist — Method

From https://www.hpl.hp.com/techreports/2003/HPL-2003-148.pdf The iterator ematchlist matches a list of terms t to a list of E-nodes by first finding all substitutions that match the first term to the first E-node, and then extending each such substitution in all possible ways that match the remaining terms to the remaining E-nodes. The base case of this recursion is the empty list, which requires no extension to the substitution; the other case relies on Match to find the substitutions that match the first term to the first E-node.

Signatures

ematchlist(e::Main.Metatheory.EGraphs.EGraph, t::Array{Any,1}, v::Array{Int64,1}, sub::Base.ImmutableDict{Any,Tuple{Main.Metatheory.EGraphs.EClass,Any}}) -> Array{Base.ImmutableDict{Any,Tuple{Main.Metatheory.EGraphs.EClass,Any}},1}

Methods

ematchlist(e, t, v, sub)

defined at /home/runner/work/Metatheory.jl/Metatheory.jl/src/EGraphs/ematch.jl:25.

Main.Metatheory.EGraphs.eval_types_in_assertions — Method

When creating a theory, type assertions in the left hand contain symbols. We want to replace the type symbols with the real type values, to fully support the subtyping mechanism during pattern matching.

Signatures

eval_types_in_assertions(x::Any, mod::Module) -> Any

Methods

eval_types_in_assertions(x, mod)

defined at /home/runner/work/Metatheory.jl/Metatheory.jl/src/EGraphs/theory_compiler.jl:31.

Main.Metatheory.EGraphs.extract! — Method

Given an ExtractionAnalysis, extract the expression with the smallest computed cost from an EGraph

Signatures

extract!(G::Main.Metatheory.EGraphs.EGraph, extran::Main.Metatheory.EGraphs.ExtractionAnalysis) -> Any

Methods

extract!(G, extran)

defined at /home/runner/work/Metatheory.jl/Metatheory.jl/src/EGraphs/extraction.jl:75.

Main.Metatheory.EGraphs.genrhsfun — Method

Generates a tuple containing the list of formal parameters (Symbols) and the RuntimeGeneratedFunction corresponding to the right hand side of a :dynamic Rule.

Signatures

genrhsfun(rule::Main.Metatheory.Rule, mod::Module) -> Tuple{Array{Symbol,1},RuntimeGeneratedFunctions.RuntimeGeneratedFunction{_A,_B,_C,_D} where _D where _C where _B where _A}

Methods

genrhsfun(rule, mod)

defined at /home/runner/work/Metatheory.jl/Metatheory.jl/src/EGraphs/theory_compiler.jl:11.

Main.Metatheory.EGraphs.isenode — Method

Check if an expr is an enode ⟺ all args are e-classes

Signatures

isenode(e::Expr) -> Bool

Methods

isenode(e)

defined at /home/runner/work/Metatheory.jl/Metatheory.jl/src/EGraphs/enode.jl:11.

Metatheory.EGraphs.Schedulers

Main.Metatheory.EGraphs.Schedulers.AbstractScheduler — Type

abstract type AbstractScheduler

Represents a rule scheduler for the equality saturation process

Fields

Main.Metatheory.EGraphs.Schedulers.BackoffScheduler — Type

struct BackoffScheduler <: Main.Metatheory.EGraphs.Schedulers.AbstractScheduler

A Rewrite Scheduler that implements exponential rule backoff. For each rewrite, there exists a configurable initial match limit. If a rewrite search yield more than this limit, then we ban this rule for number of iterations, double its limit, and double the time it will be banned next time.

This seems effective at preventing explosive rules like associativity from taking an unfair amount of resources.

Fields

data::Dict{Main.Metatheory.Rule,Main.Metatheory.EGraphs.Schedulers.BackoffSchedulerEntry}

G::Main.Metatheory.EGraphs.EGraph

theory::Array{Main.Metatheory.Rule,1}

Main.Metatheory.EGraphs.Schedulers.SimpleScheduler — Type

struct SimpleScheduler <: Main.Metatheory.EGraphs.Schedulers.AbstractScheduler

A simple Rewrite Scheduler that applies every rule every time

Fields

Main.Metatheory.EGraphs.Schedulers.cansaturate — Method

Should return true if the e-graph can be said to be saturated

Signatures

cansaturate(s::Main.Metatheory.EGraphs.Schedulers.AbstractScheduler) -> Bool

Methods

cansaturate(s)

defined at /home/runner/work/Metatheory.jl/Metatheory.jl/src/EGraphs/Schedulers/abstractscheduler.jl:12.

cansaturate(s)

defined at /home/runner/work/Metatheory.jl/Metatheory.jl/src/EGraphs/Schedulers/backoffscheduler.jl:43.

cansaturate(s)

defined at /home/runner/work/Metatheory.jl/Metatheory.jl/src/EGraphs/Schedulers/simplescheduler.jl:6.

Main.Metatheory.EGraphs.Schedulers.readstep! — Method

This function is called before pattern matching on the e-graph

Signatures

readstep!(s::Main.Metatheory.EGraphs.Schedulers.AbstractScheduler)

Methods

readstep!(s)

defined at /home/runner/work/Metatheory.jl/Metatheory.jl/src/EGraphs/Schedulers/abstractscheduler.jl:22.

readstep!(s)

defined at /home/runner/work/Metatheory.jl/Metatheory.jl/src/EGraphs/Schedulers/backoffscheduler.jl:45.

readstep!(s)

defined at /home/runner/work/Metatheory.jl/Metatheory.jl/src/EGraphs/Schedulers/simplescheduler.jl:13.

Main.Metatheory.EGraphs.Schedulers.shouldskip — Method

Should return true if the rule r should be skipped

Signatures

shouldskip(s::Main.Metatheory.EGraphs.Schedulers.AbstractScheduler, r::Main.Metatheory.Rule) -> Bool

Methods

shouldskip(s, r)

defined at /home/runner/work/Metatheory.jl/Metatheory.jl/src/EGraphs/Schedulers/abstractscheduler.jl:17.

shouldskip(s, r)

defined at /home/runner/work/Metatheory.jl/Metatheory.jl/src/EGraphs/Schedulers/backoffscheduler.jl:27.

shouldskip(s, r)

defined at /home/runner/work/Metatheory.jl/Metatheory.jl/src/EGraphs/Schedulers/simplescheduler.jl:7.

Main.Metatheory.EGraphs.Schedulers.writestep! — Method

This function is called after pattern matching on the e-graph

Signatures

writestep!(s::Main.Metatheory.EGraphs.Schedulers.AbstractScheduler, r::Main.Metatheory.Rule)

Methods

writestep!(s, r)

defined at /home/runner/work/Metatheory.jl/Metatheory.jl/src/EGraphs/Schedulers/abstractscheduler.jl:27.

writestep!(s, rule)

defined at /home/runner/work/Metatheory.jl/Metatheory.jl/src/EGraphs/Schedulers/backoffscheduler.jl:61.

writestep!(s, r)

defined at /home/runner/work/Metatheory.jl/Metatheory.jl/src/EGraphs/Schedulers/simplescheduler.jl:14.

Metatheory.Util

Main.Metatheory.Util — Module

Definitions of various utility functions for metaprogramming

Imports

Base

Base.Meta

Core

DocStringExtensions

Main.Metatheory.Util.amp — Method

Add a & expression

Signatures

amp(v::Any) -> Expr

Methods

amp(v)

defined at /home/runner/work/Metatheory.jl/Metatheory.jl/src/Util/Util.jl:15.

Main.Metatheory.Util.bf_walk! — Method

Breadth First Walk (Tree Prewalk) on expressions mutates expression in-place.

Signatures

bf_walk!(f::Any, e::Any, f_args::Vararg{Any,N} where N; skip, skip_call) -> Any

Methods

bf_walk!(f, e, f_args; skip, skip_call)

defined at /home/runner/work/Metatheory.jl/Metatheory.jl/src/Util/walks.jl:45.

Main.Metatheory.Util.bf_walk — Method

Breadth First Walk (Tree Prewalk) on expressions. Does not mutate expressions.

Signatures

bf_walk(f::Any, e::Any, f_args::Vararg{Any,N} where N; skip, skip_call) -> Any

Methods

bf_walk(f, e, f_args; skip, skip_call)

defined at /home/runner/work/Metatheory.jl/Metatheory.jl/src/Util/walks.jl:65.

Main.Metatheory.Util.binarize! — Method

HARD FIX of n-arity of operators in Expr trees

Signatures

binarize!(e::Any, op::Symbol) -> Any

Methods

binarize!(e, op)

defined at /home/runner/work/Metatheory.jl/Metatheory.jl/src/Util/cleaning.jl:11.

Main.Metatheory.Util.block — Method

Make a block expression from an array of exprs

Signatures

block(vs::Vararg{Any,N} where N) -> Any

Methods

block(vs)

defined at /home/runner/work/Metatheory.jl/Metatheory.jl/src/Util/Util.jl:13.

Main.Metatheory.Util.cleanast — Method

Binarize n-ary operators (+ and *) and call rmlines

Signatures

cleanast(ex::Any) -> Any

Methods

cleanast(ex)

defined at /home/runner/work/Metatheory.jl/Metatheory.jl/src/Util/cleaning.jl:22.

Main.Metatheory.Util.df_walk! — Method

Depth First Walk (Tree Postwalk) on expressions, mutates expression in-place.

Signatures

df_walk!(f::Any, e::Any, f_args::Vararg{Any,N} where N; skip, skip_call) -> Any

Methods

df_walk!(f, e, f_args; skip, skip_call)

defined at /home/runner/work/Metatheory.jl/Metatheory.jl/src/Util/walks.jl:4.

Main.Metatheory.Util.df_walk — Method

Depth First Walk (Tree Postwalk) on expressions. Does not mutate expressions.

Signatures

df_walk(f::Any, e::Any, f_args::Vararg{Any,N} where N; skip, skip_call) -> Any

Methods

df_walk(f, e, f_args; skip, skip_call)

defined at /home/runner/work/Metatheory.jl/Metatheory.jl/src/Util/walks.jl:22.

Main.Metatheory.Util.dollar — Method

Add a dollar expression

Signatures

dollar(v::Any) -> Expr

Methods

dollar(v)

defined at /home/runner/work/Metatheory.jl/Metatheory.jl/src/Util/Util.jl:11.

Main.Metatheory.Util.normalize — Method

Iterates a function f on datum until a fixed point is reached where f(x) == x

Signatures

normalize(f::Any, datum::Any, fargs::Vararg{Any,N} where N; callback) -> Any

Methods

normalize(f, datum, fargs; callback)

defined at /home/runner/work/Metatheory.jl/Metatheory.jl/src/Util/fixpoint.jl:4.

Main.Metatheory.Util.normalize_nocycle — Method

Like normalize but keeps a vector of hashes to detect cycles, returns the current datum when a cycle is detected

Signatures

normalize_nocycle(f::Any, datum::Any, fargs::Vararg{Any,N} where N; callback) -> Any

Methods

normalize_nocycle(f, datum, fargs; callback)

defined at /home/runner/work/Metatheory.jl/Metatheory.jl/src/Util/fixpoint.jl:20.

Main.Metatheory.Util.rmlines — Method

Remove LineNumberNode from quoted blocks of code

Signatures

rmlines(e::Expr) -> Any

Methods

rmlines(e)

defined at /home/runner/work/Metatheory.jl/Metatheory.jl/src/Util/cleaning.jl:4.