API Documentation

Metatheory

Main.Metatheory.Rule — Type

Structure representing rewrite rules.

Main.Metatheory.compile_theory — Method

Compile a theory to a closure that does the pattern matching job Returns a RuntimeGeneratedFunction, which does not use eval and is as fast as a regular Julia anonymous function 🔥

Main.Metatheory.@compile_theory — Macro

Compile a theory at runtime to a closure that does the pattern matching job

Metatheory.EGraphs

Main.Metatheory.EGraphs.RhsFunCache — Type

Type representing a cache of RuntimeGeneratedFunctions corresponding to right hand sides of dynamic rules

Main.Metatheory.EGraphs.eval_types_in_assertions — Method

When creating a theory, type assertions in the left hand contain symbols. We want to replace the type symbols with the real type values, to fully support the subtyping mechanism during pattern matching.

Main.Metatheory.EGraphs.genrhsfun — Method

Generates a tuple containing the list of formal parameters (Symbols) and the RuntimeGeneratedFunction corresponding to the right hand side of a :dynamic Rule.

Metatheory.EGraphs.Schedulers

Main.Metatheory.EGraphs.Schedulers.AbstractScheduler — Type

Represents a rule scheduler for the equality saturation process

Main.Metatheory.EGraphs.Schedulers.BackoffScheduler — Type

A Rewrite Scheduler that implements exponential rule backoff. For each rewrite, there exists a configurable initial match limit. If a rewrite search yield more than this limit, then we ban this rule for number of iterations, double its limit, and double the time it will be banned next time.

This seems effective at preventing explosive rules like associativity from taking an unfair amount of resources.

Main.Metatheory.EGraphs.Schedulers.SimpleScheduler — Type

A simple Rewrite Scheduler that applies every rule every time

Main.Metatheory.EGraphs.Schedulers.cansaturate — Method

Should return true if the e-graph can be said to be saturated

Main.Metatheory.EGraphs.Schedulers.readstep! — Method

This function is called before pattern matching on the e-graph

Main.Metatheory.EGraphs.Schedulers.shouldskip — Method

Should return true if the rule r should be skipped

Main.Metatheory.EGraphs.Schedulers.writestep! — Method

This function is called after pattern matching on the e-graph

Metatheory.Util

Main.Metatheory.Util — Module

Definitions of various utility functions for metaprogramming

Main.Metatheory.Util.amp — Method

Add a & expression

Main.Metatheory.Util.bf_walk! — Method

Breadth First Walk (Tree Prewalk) on expressions mutates expression in-place.

Main.Metatheory.Util.bf_walk — Method

Breadth First Walk (Tree Prewalk) on expressions. Does not mutate expressions.

Main.Metatheory.Util.binarize! — Method

HARD FIX of n-arity of operators in Expr trees

Main.Metatheory.Util.block — Method

Make a block expression from an array of exprs

Main.Metatheory.Util.cleanast — Method

Binarize n-ary operators (+ and *) and call rmlines

Main.Metatheory.Util.df_walk! — Method

Depth First Walk (Tree Postwalk) on expressions, mutates expression in-place.

Main.Metatheory.Util.df_walk — Method

Depth First Walk (Tree Postwalk) on expressions. Does not mutate expressions.

Main.Metatheory.Util.dollar — Method

Add a dollar expression

Main.Metatheory.Util.normalize — Method

Iterates a function f on datum until a fixed point is reached where f(x) == x

Main.Metatheory.Util.normalize_nocycle — Method

Like normalize but keeps a vector of hashes to detect cycles, returns the current datum when a cycle is detected

Main.Metatheory.Util.rmlines — Method

Remove LineNumberNode from quoted blocks of code