Metatheory

Metatheory.compile_theory — Method

Compile a theory to a closure that does the pattern matching job Returns a RuntimeGeneratedFunction, which does not use eval and is as fast as a regular Julia anonymous function 🔥

Metatheory.@compile_theory — Macro

Compile a theory at runtime to a closure that does the pattern matching job