Scheduling for E-Graph rewriting

Metatheory.EGraphs.Schedulers.AbstractScheduler — Type

abstract type AbstractScheduler

Represents a rule scheduler for the equality saturation process

Fields

Metatheory.EGraphs.Schedulers.BackoffScheduler — Type

mutable struct BackoffScheduler <: Metatheory.EGraphs.Schedulers.AbstractScheduler

A Rewrite Scheduler that implements exponential rule backoff. For each rewrite, there exists a configurable initial match limit. If a rewrite search yield more than this limit, then we ban this rule for number of iterations, double its limit, and double the time it will be banned next time.

This seems effective at preventing explosive rules like associativity from taking an unfair amount of resources.

Fields

data::IdDict{Rule, Metatheory.EGraphs.Schedulers.BackoffSchedulerEntry}

G::EGraph

theory::Vector{var"#s213"} where var"#s213"<:Rule

curr_iter::Int64

Metatheory.EGraphs.Schedulers.ScoredScheduler — Type

mutable struct ScoredScheduler <: Metatheory.EGraphs.Schedulers.AbstractScheduler

A Rewrite Scheduler that implements exponential rule backoff. For each rewrite, there exists a configurable initial match limit. If a rewrite search yield more than this limit, then we ban this rule for number of iterations, double its limit, and double the time it will be banned next time.

This seems effective at preventing explosive rules like associativity from taking an unfair amount of resources.

Fields

data::IdDict{Rule, Metatheory.EGraphs.Schedulers.ScoredSchedulerEntry}

G::EGraph

theory::Vector{var"#s213"} where var"#s213"<:Rule

curr_iter::Int64

Metatheory.EGraphs.Schedulers.SimpleScheduler — Type

struct SimpleScheduler <: Metatheory.EGraphs.Schedulers.AbstractScheduler

A simple Rewrite Scheduler that applies every rule every time

Fields

Metatheory.EGraphs.Schedulers.cansaturate — Function

Should return true if the e-graph can be said to be saturated

cansaturate(s::AbstractScheduler)

Signatures

Methods

cansaturate(s)

defined at /home/runner/work/Metatheory.jl/Metatheory.jl/src/EGraphs/Schedulers/backoffscheduler.jl:45.

cansaturate(s)

defined at /home/runner/work/Metatheory.jl/Metatheory.jl/src/EGraphs/Schedulers/scoredscheduler.jl:84.

cansaturate(s)

defined at /home/runner/work/Metatheory.jl/Metatheory.jl/src/EGraphs/Schedulers/simplescheduler.jl:6.

Metatheory.EGraphs.Schedulers.cansearch — Function

Should return false if the rule r should be skipped

cansearch(s::AbstractScheduler, r::Rule)

Signatures

Methods

cansearch(s, r)

defined at /home/runner/work/Metatheory.jl/Metatheory.jl/src/EGraphs/Schedulers/backoffscheduler.jl:25.

cansearch(s, r)

defined at /home/runner/work/Metatheory.jl/Metatheory.jl/src/EGraphs/Schedulers/scoredscheduler.jl:26.

cansearch(s, r)

defined at /home/runner/work/Metatheory.jl/Metatheory.jl/src/EGraphs/Schedulers/simplescheduler.jl:7.

Metatheory.EGraphs.Schedulers.inform! — Function

This function is called after pattern matching on the e-graph, informs the scheduler about the yielded matches. Returns false if the matches should not be yielded and ignored.

inform!(s::AbstractScheduler, iteration, r::Rule, matches)

Signatures

Methods

inform!(s, rule, matches)

defined at /home/runner/work/Metatheory.jl/Metatheory.jl/src/EGraphs/Schedulers/backoffscheduler.jl:48.

inform!(s, rule, matches)

defined at /home/runner/work/Metatheory.jl/Metatheory.jl/src/EGraphs/Schedulers/scoredscheduler.jl:87.

inform!(s, r, matches)

defined at /home/runner/work/Metatheory.jl/Metatheory.jl/src/EGraphs/Schedulers/simplescheduler.jl:13.